Added Thing ListTop(List const L);    // thse can return NULL if the list size Thing ListEnd(List const L);    // is 0 for the internal lib workings, needed by the Hash table, but not for the Api. ListTop skips over the dummyhead which is of Type Special
Careful with HashGet as it can return NULL;
DO NOT DO PASS THE RETURN VALUE OF HashGet BLINDLY TO ANOTHER FUNCTION:
if (IntWord(HashGet(H, key)) ... )
it will just lead to sadness.

Hash Table tested somewhat.  Problem being if there is collisions it will only
return the List->Top.  Have to think of a solution for this.

Need a Better Hash Function